# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Maryam Abacha American University of Nigeria - University Management Portal API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /students:
    get:
      operationId: listStudents
      tags: [students]
      summary: List all students (admin/lecturer)
      parameters:
        - name: department
          in: query
          schema:
            type: string
        - name: level
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Students
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/StudentProfile"
  /students/{id}:
    get:
      operationId: getStudent
      tags: [students]
      summary: Get student profile
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Student profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StudentProfile"
    put:
      operationId: updateStudent
      tags: [students]
      summary: Update student profile
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateStudentBody"
      responses:
        "200":
          description: Updated student
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StudentProfile"
  /students/{id}/cgpa:
    get:
      operationId: getStudentCgpa
      tags: [students]
      summary: Calculate and get student CGPA
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: CGPA data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CgpaResponse"
...
components:
  schemas:
...
    StudentProfile:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        name:
          type: string
        email:
          type: string
        matricNumber:
          type: string
        department:
          type: string
        faculty:
          type: string
        level:
          type: string
        cgpa:
          type: number
          nullable: true
        enrollmentYear:
          type: string
        createdAt:
          type: string
      required: [id, userId, name, email, matricNumber, department, faculty, level, enrollmentYear, createdAt]
...
    UpdateStudentBody:
      type: object
      properties:
        department:
          type: string
        faculty:
          type: string
        level:
          type: string
      required: []
    CgpaResponse:
      type: object
      properties:
        studentId:
          type: integer
        cgpa:
          type: number
        totalUnitsAttempted:
          type: integer
        totalQualityPoints:
          type: number
        semesterBreakdown:
          type: array
          items:
            $ref: "#/components/schemas/SemesterCgpa"
      required: [studentId, cgpa, totalUnitsAttempted, totalQualityPoints, semesterBreakdown]
    SemesterCgpa:
      type: object
      properties:
        semester:
          type: string
        academicYear:
          type: string
        gpa:
          type: number
        unitsAttempted:
          type: integer
        qualityPoints:
          type: number
      required: [semester, academicYear, gpa, unitsAttempted, qualityPoints]
...